third_party/WebKit/LayoutTests/plugins/overlay-scrollbar-mouse-capture-expected.txt


This tests whether scrolling still works correctly when an overlay scrollbar is over a plugin. Clicking on the overlay scrollbar should cause it to activate and capture input.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS eventHistory.length is 2
PASS eventHistory[0] is "plugin.mousedown"
PASS eventHistory[1] is "plugin.mouseup"
PASS window.scrollY > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
